+++ Assets/Member/Dimu/-----Scripts/Interact/Item.cs
@@ -12,7 +12,7 @@ public class Item : ScriptableObject
     [SerializeField] private string itemDescription;
     [SerializeField] private Sprite itemSprite;
     [SerializeField] private Material itemMaterial;
-    [SerializeField] int cleanAmount;
+    [SerializeField] float cleanAmount;
     [SerializeField] int eatAmount;
     //[Header("Combination")]
     //[SerializeField] private ingredientNeeded[] necessaryIngredient;
@@ -103,7 +103,7 @@ public class Item : ScriptableObject
     //public OutItemType OutItemType => outItemType;
     //public Sprite ItemSprite => itemSprite;
     //public Material ItemMaterial => itemMaterial;
-    public int CleanAmount => cleanAmount;
+    public float CleanAmount => cleanAmount;
     public int EatAmount => eatAmount;
     //public ingredientNeeded[] NnecessaryIngredient => necessaryIngredient;
     public DropItem[] DropItems => dropItems;
